refactored SingleSong page
modularized buttons
