+++ react-app/src/components/SongFolders/SingleSong/Comments/SingleComment.css
@@ -1,113 +1,120 @@
 .comment-card {
-    padding: 10px 10px 10px 0;
-    justify-content: space-between;
+  padding: 10px 10px 10px 0;
+  justify-content: space-between;
 }
 
 .comment-body {
-    width: 100%;
-    align-items: center;
-    gap: 10px;
+  width: 100%;
+  align-items: center;
+  gap: 10px;
+}
+
+.comment-avatar .avatar {
+  display: block;
+  width: 40px;
+  height: 40px;
 }
 
 .comment-info,
 .comment-info *,
 .comment-meta {
-    color: #999;
+  color: #999;
 }
 
 .commenter-name,
 .comment-timestamp {
-    font-weight: 900;
+  font-weight: 900;
 }
 
 .comment-text {
-    width: 100%;
+  width: 100%;
 }
 
 .comment-content {
-    width: 100%;
+  width: 100%;
 }
 
 .content-edit {
-    width: 100%;
-    padding: 5px;
+  width: 100%;
+  padding: 5px;
 }
 
 .content-field {
-    background-color: var(--color-border-light);
-    width: 100%;
-    border: none;
-    border-bottom: 1px solid var(--color-border);
+  background-color: var(--color-border-light);
+  width: 100%;
+  border: none;
+  border-bottom: 1px solid var(--color-border);
 }
 
 .content-field:focus {
-    outline: none;
+  outline: none;
 }
 
 .comment-meta {
-    align-items: flex-end;
-    min-width: 150px;
+  align-items: flex-end;
+  min-width: 150px;
+  gap: 5px;
 }
 
 .button {
-    align-items: flex-end;
-    border: 1px solid #e5e5e5;
-    border-radius: 3px;
-    cursor: pointer;
-    color: #333;
-    font-size: 11px;
-    padding: 2px 9px 2px 8px;
-    height: 22px;
-    line-height: 16px;
-    gap: 4px;
+  align-items: flex-end;
+  border: 1px solid #e5e5e5;
+  border-radius: 3px;
+  cursor: pointer;
+  color: #333;
+  font-size: 11px;
+  padding: 2px 9px 2px 8px;
+  height: 22px;
+  line-height: 16px;
+  gap: 4px;
 }
 
 .button:hover {
-    border-color: #999;
+  border-color: #999;
 }
 
 .edit-icon {
-    position: relative;
-    bottom: -1px;
+  position: relative;
+  bottom: -1px;
 }
 
 .reply-icon {
-    background-image: url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAA8AAAAPCAQAAACR313BAAAAGXRFWHRTb2Z0d2FyZQBBZG9iZSBJbWFnZVJlYWR5ccllPAAAALRJREFUGNNjYCAexMnhkYyxTHiJUzIqIeFX/n+sUhHMEZMT/ucBIRbJMKHw3Un/s8Ew7H/Y27CTYe1h0gi9+5L+Z8Bh2v+U/3H/w74Eu0Kl30xueZ/2HxWm/o/4GqQIlv7P+3Py7CfJ/0Hw//7Ph2+cOXhl0ouMP3HzoPr/s/zP23Ar6W8C0Gn/uf+r/bf/X/liX+MhJAf+Dzx7Ju0XEl/kVTmKD/7rP9qDwudG8+J/SQaqAgBt9GUu05AvWgAAAABJRU5ErkJggg==);
-    width: 16px;
-    height: 16px;
+  background-image: url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAA8AAAAPCAQAAACR313BAAAAGXRFWHRTb2Z0d2FyZQBBZG9iZSBJbWFnZVJlYWR5ccllPAAAALRJREFUGNNjYCAexMnhkYyxTHiJUzIqIeFX/n+sUhHMEZMT/ucBIRbJMKHw3Un/s8Ew7H/Y27CTYe1h0gi9+5L+Z8Bh2v+U/3H/w74Eu0Kl30xueZ/2HxWm/o/4GqQIlv7P+3Py7CfJ/0Hw//7Ph2+cOXhl0ouMP3HzoPr/s/zP23Ar6W8C0Gn/uf+r/bf/X/liX+MhJAf+Dzx7Ju0XEl/kVTmKD/7rP9qDwudG8+J/SQaqAgBt9GUu05AvWgAAAABJRU5ErkJggg==);
+  width: 16px;
+  height: 16px;
 }
 
 .delete-icon {
-    background-image: url(https://a-v2.sndcdn.com/assets/images/delete-d90bf5e4.svg);
-    width: 16px;
-    height: 16px;
+  background-image: url(https://a-v2.sndcdn.com/assets/images/delete-d90bf5e4.svg);
+  width: 16px;
+  height: 16px;
 }
 
 .delete-confirm .button:hover,
 .yes-delete.button:hover {
-    border-color: #e30;
+  border-color: #e30;
 }
 
 .confirm-delete {
-    position: relative;
+  position: relative;
 }
 
 .confirm-delete-form {
-    position: absolute;
-    width: 200px;
-    background-color: white;
-    z-index: 5;
-    top: 5px;
-    right: 0px;
-    border: 1px solid #e5e5e5;
-    padding: 10px;
+  position: absolute;
+  width: 200px;
+  background-color: white;
+  z-index: 5;
+  top: 5px;
+  right: 0px;
+  border: 1px solid #e5e5e5;
+  padding: 10px;
 }
 
 .confirm-delete-form p {
-    color: #333;
-    margin-bottom: 5px;
+  color: #333;
+  margin-bottom: 5px;
 }
 
 .confirm-delete-buttons {
-    justify-content: flex-end;
+  justify-content: flex-end;
 }
