+++ react-app/src/components/SongFolders/SingleSong/SingleSong.css
@@ -60,6 +60,7 @@
 
 .single-song-secondary {
   gap: 30px;
+  padding-top: 20px;
 }
 
 .single-song-main {
@@ -106,12 +107,7 @@
   width: 60px;
 }
 
-.song_mainfeed_top {
-  margin-top: 20px;
-  height: 90px;
-}
-
-.comment-section {
+.comment-two-columns {
   gap: 20px;
 }
 
@@ -153,13 +149,29 @@
   outline: none;
 }
 
-.user-badge .avatar {
-  height: 120px;
+.button-group-menu {
+  background-color: white;
+  box-shadow: 0px 0px 4px var(--color-border);
   width: 120px;
 }
 
-.song-details {
-  width: 100%;
+.button-group-menu li {
+  height: 24px;
+  align-items: center;
+  padding-left: 10px;
+  font-size: 14px;
+}
+
+.dropdown-clipboard {
+  position: absolute;
+  background-color: white;
+  box-shadow: 0px 0px 4px black;
+  padding: 2px 8px;
+}
+
+.user-badge .avatar {
+  height: 120px;
+  width: 120px;
 }
 
 .song-description {
@@ -182,7 +194,7 @@
 .comments-count {
   padding-bottom: 7px;
   border-bottom: 1px solid #f2f2f2;
-  width: 100%;
+  /* width: 100%; */
   align-items: flex-end;
 }
 
@@ -192,22 +204,22 @@
   color: #999;
 }
 
-.button-group-menu {
-  background-color: white;
-  box-shadow: 0px 0px 4px var(--color-border);
-  width: 120px;
-}
-
-.button-group-menu li {
-  height: 24px;
+.no-comments {
+  width: 100%;
+  justify-content: center;
   align-items: center;
-  padding-left: 10px;
-  font-size: 14px;
+  height: 500px;
+  gap: 10px;
 }
 
-.dropdown-clipboard {
-  position: absolute;
-  background-color: white;
-  box-shadow: 0px 0px 4px black;
-  padding: 2px 8px;
+.no-comments h3 {
+  font-size: 18px;
+  font-weight: 100;
+  color: var(--color-border-dark);
+}
+
+.no-comments h4 {
+  font-size: 16px;
+  font-weight: 100;
+  color: var(--color-info);
 }
